requirements.txt
beautifulsoup4==4.12.3
fake-useragent==1.5.1
requests==2.32.3
python-dotenv~=1.0.1
psycopg==3.2.3
psycopg-binary==3.2.3
